Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -11,7 +11,7 @@ int main() {
         scanf("%d %d",&s[i].a,&s[i].b);
     }
     for(int i=0;i<n;i++){
-        printf("Rectangle %d: Area = %.2f, Perimeter = %.2f",i,s[i].a*s[i].b,2*(s[i].a+a[i].b));
+        printf("Rectangle %d: Area = %.2f, Perimeter = %.2f",i,float(s[i].a*s[i]).b,2*(s[i].a+a[i].b));
     }
     
     return 0;
